Let's make sure EnsoContext is garbage collected after each ContextUtils usage
